ReplicatedPG,rados: add CEPH_OSD_[COPY_FROM]_IGNORE_RMED_SNAP
When attempting to promote a particular clone, we need to be able to able to
handle the posibility that the snap we are using to specify the clone may have
been removed.  Also, the cloneid used by the backing pool may not be the same
as the clone id used in the cache pool in the event that the cache pool's
cloneid has been removed before the clone in question has been flushed.  Thus,
we will promote using a snap which currently maps to the clone we want.  The
backend will do 1 of 2 things:
1) If the clone still exists, our requested snap will map to that clone
2) If the clone has been trimmed, we may end up reading from another
subsequent object.
We handle case 2) by, in finish_promote, verifying that our current pool info
indicates that the clone should still exist.

Fixes: #7858
Signed-off-by: Samuel Just <sam.just@inktank.com>